requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
pymongo==3.12.0
dnspython==2.1.0
bcrypt==3.2.0
PyJWT==2.1.0
python-dotenv==0.19.0
gunicorn==20.1.0